Fase 1: subida real de datasets a MinIO con inferencia de esquema
POST /datasets pasa de recibir metadata en JSON a recibir el archivo por
multipart: lo guarda en MinIO, infiere el esquema con DuckDB (columna,
tipo, nulos, filas) y persiste la ficha en Postgres.

- app/storage.py: cliente S3 (boto3) contra MinIO, bucket auto-creado
- app/schema_inference.py: inferencia CSV/Parquet con DuckDB
- migracion 0002: size_bytes y row_count_estimate a bigint
- script.py.mako: plantilla de migraciones que pasa ruff (I001, sintaxis
  moderna de typing) para no arrastrar el lint en cada migracion futura
- traefik 3.6 en compose: 3.1.7 pedia API 1.24 y el daemon exige 1.44+,
  por lo que no descubria ningun router
- frontend: selector de archivo real y tabla del esquema inferido
- CI: servicio MinIO junto a Postgres

Los tests de subida corren contra un MinIO real, no contra un mock de S3:
un mock confirma la llamada, no que el objeto quede guardado y se pueda
recuperar. 13 tests en verde (antes 5), ruff y mypy limpios, y flujo
extremo a extremo verificado por Traefik con un CSV real.

+++ backend/app/models.py
@@ -10,7 +10,7 @@ import enum
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Enum, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, Enum, ForeignKey, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
@@ -57,8 +57,12 @@ class Dataset(Base):
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     source_uri: Mapped[str | None] = mapped_column(String(1024), nullable=True)
     format: Mapped[DatasetFormat] = mapped_column(Enum(DatasetFormat, name="dataset_format"), nullable=False)
-    size_bytes: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    row_count_estimate: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # BigInteger y no Integer: int4 topea en 2.147.483.647, o sea ~2,1 GB de
+    # tamano y ~2.100 millones de filas. Para una herramienta cuyo proposito es
+    # justamente masticar datasets grandes, ese techo se alcanza con un solo
+    # archivo y el desbordamiento aparece recien al insertar, en produccion.
+    size_bytes: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    row_count_estimate: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     # Esquema inferido: [{"name": "col", "dtype": "float64", "nullable": true}, ...]
     inferred_schema: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
     version: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
